Refactor the main view body
