+++ CodeBreaker/CodeBreaker/UI/UIExtensions.swift
@@ -1,0 +1,33 @@
+//
+//  UIExtensions.swift
+//  CodeBreaker
+//
+//  Created by Alexander Ostrovsky on 9/2/2026.
+//
+
+import SwiftUI
+
+extension Animation {
+    static let codeBreaker = Animation.easeInOut(duration: 3)
+    static let guess = Animation.codeBreaker
+    static let restart = Animation.codeBreaker
+}
+
+extension AnyTransition {
+    static let pegChooser = AnyTransition.offset(y: 200)
+    static func attempt(_ isOver: Bool) -> AnyTransition {
+        AnyTransition.asymmetric(insertion: isOver ? .opacity : .move(edge: .top), removal: .move(edge: .trailing))
+    }
+}
+
+extension View {
+    func flexibleSystemFont(minimum: CGFloat = 8, maximum: CGFloat = 80) -> some View {
+        font(.system(size: maximum)).minimumScaleFactor(minimum / maximum)
+    }
+}
+
+extension Color {
+    static func gray(_ brightness: CGFloat) -> Color {
+        .init(hue: 148/360, saturation: 0, brightness: brightness)
+    }
+}
